REB3-8088. Fixed circular dependency.
